// src/generated/client/models/KnotType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Metamind API
 * Brain spec for Metamind.
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum KnotType {
    TEXT = 'TEXT',
    SCRIPT = 'SCRIPT',
    IMAGE = 'IMAGE',
    EMAIL = 'EMAIL'
}

export function KnotTypeFromJSON(json: any): KnotType {
    return KnotTypeFromJSONTyped(json, false);
}

export function KnotTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): KnotType {
    return json as KnotType;
}

export function KnotTypeToJSON(value?: KnotType | null): any {
    return value as any;
}