// src/generated/client/models/Story.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Metamind API
 * Brain spec for Metamind.
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Story
 */
export interface Story {
    /**
     * Story id.
     * @type {string}
     * @memberof Story
     */
    readonly id?: string;
    /**
     * default hint for the story
     * @type {string}
     * @memberof Story
     */
    dafaultHint?: string;
    /**
     * Story name.
     * @type {string}
     * @memberof Story
     */
    name: string;
    /**
     * Story locale.
     * @type {string}
     * @memberof Story
     */
    locale: string;
    /**
     * List of story global quick responses.
     * @type {Array<string>}
     * @memberof Story
     */
    readonly quickResponses?: Array<string>;
    /**
     * Creation time
     * @type {Date}
     * @memberof Story
     */
    readonly createdAt?: Date;
    /**
     * Last modification time
     * @type {Date}
     * @memberof Story
     */
    readonly modifiedAt?: Date;
}

export function StoryFromJSON(json: any): Story {
    return StoryFromJSONTyped(json, false);
}

export function StoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): Story {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'dafaultHint': !exists(json, 'dafaultHint') ? undefined : json['dafaultHint'],
        'name': json['name'],
        'locale': json['locale'],
        'quickResponses': !exists(json, 'quickResponses') ? undefined : json['quickResponses'],
        'createdAt': !exists(json, 'createdAt') ? undefined : (new Date(json['createdAt'])),
        'modifiedAt': !exists(json, 'modifiedAt') ? undefined : (new Date(json['modifiedAt'])),
    };
}

export function StoryToJSON(value?: Story | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'dafaultHint': value.dafaultHint,
        'name': value.name,
        'locale': value.locale,
    };
}